Merge in order to allow a fast-forward to the release tag

+++ tests/unit/src/TimeValueSamples.cpp
@@ -1,12 +1,12 @@
 /**
 * Copyright 2016 BitTorrent Inc.
-* 
+*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
-* 
+*
 *    http://www.apache.org/licenses/LICENSE-2.0
-* 
+*
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -57,13 +57,13 @@ TEST(TimeValueSamples, erasesOlderSamples) {
 }
 
 TEST(TimeValueSamples, oldestNewest) {
-    TimeValueSamples<uintmax_t> timeValues{3};
+    TimeValueSamples<int> timeValues{3};
     EXPECT_EQ(timeValues.size(), 0);
 
     timeValues.insert(std::chrono::steady_clock::time_point(1s), 1000);
     timeValues.insert(std::chrono::steady_clock::time_point(2s), 2000);
     timeValues.insert(std::chrono::steady_clock::time_point(3s), 3000);
 
-    EXPECT_EQ(timeValues.oldest(), std::make_pair(std::chrono::steady_clock::time_point(1s), 1000ul));
-    EXPECT_EQ(timeValues.newest(), std::make_pair(std::chrono::steady_clock::time_point(3s), 3000ul));
+    EXPECT_EQ(timeValues.oldest(), std::make_pair(std::chrono::steady_clock::time_point(1s), 1000));
+    EXPECT_EQ(timeValues.newest(), std::make_pair(std::chrono::steady_clock::time_point(3s), 3000));
 }
